+++ martin_binance/backtest/optimizer.py
@@ -6,7 +6,7 @@ Searches for optimal parameters for a strategy under given conditions
 __author__ = "Jerry Fedorenko"
 __copyright__ = "Copyright © 2024 Jerry Fedorenko aka VM"
 __license__ = "MIT"
-__version__ = "2.1.0"
+__version__ = "2.1.0rc16"
 __maintainer__ = "Jerry Fedorenko"
 __contact__ = "https://github.com/DogsTailFarmer"
 
@@ -15,21 +15,17 @@ import importlib.util as iu
 from decimal import Decimal
 import optuna
 import asyncio
+import sys
+import ujson as json
+from pathlib import Path
+import stat
+import logging
 
+OPTIMIZER = Path(__file__).absolute()
+OPTIMIZER.chmod(OPTIMIZER.stat().st_mode | stat.S_IEXEC)
 PARAMS_FLOAT = ['KBB']
-N_TRIALS = 5  # 500
-
-
-async def run_optimize(*args):
-    # Create subprocess
-    process = await asyncio.create_subprocess_exec(
-        *args,
-        # stdout must a pipe to be accessible as process.stdout
-        stdout=asyncio.subprocess.PIPE)
-    # Wait for the subprocess to finish
-    stdout, stderr = await process.communicate()
-    # Return stdout
-    return stdout.decode().strip()
+N_TRIALS = 1000
+logger = logging.getLogger('logger')
 
 
 def try_trade(mbs, **kwargs):
@@ -42,7 +38,7 @@ def try_trade(mbs, **kwargs):
     return float(mbs.session_result.get('profit', 0)) + float(mbs.session_result.get('free', 0))
 
 
-def optimize(study_name, strategy, n_trials=0, storage_name=None):
+def optimize(study_name, strategy, n_trials, storage_name=None):
     def objective(_trial):
         params = {
             'GRID_MAX_COUNT': _trial.suggest_int('GRID_MAX_COUNT', 3, 5),
@@ -58,11 +54,38 @@ def optimize(study_name, strategy, n_trials=0, storage_name=None):
         }
         return try_trade(mbs, **params)
 
-    print(f"Importing strategy from {strategy}")
-
     spec = iu.spec_from_file_location("strategy", strategy)
     mbs = iu.module_from_spec(spec)
     spec.loader.exec_module(mbs)
-    study = optuna.create_study(study_name=study_name, storage=storage_name, direction="maximize")
-    study.optimize(objective, n_trials=n_trials or N_TRIALS)
-    return study
+    _study = optuna.create_study(study_name=study_name, storage=storage_name, direction="maximize")
+    optuna.logging.set_verbosity(optuna.logging.WARNING)
+    try:
+        _study.optimize(objective, n_trials=n_trials)
+    except KeyboardInterrupt:
+        pass
+    return _study
+
+
+async def run_optimize(*args):
+    process = await asyncio.create_subprocess_exec(
+        *args,
+        stdout=asyncio.subprocess.PIPE,
+        stderr=asyncio.subprocess.PIPE
+    )
+    stdout, stderr = await process.communicate()
+    logger.info(stderr.decode().strip())
+    return stdout
+
+
+if __name__ == "__main__":
+    try:
+        study = optimize(sys.argv[1], sys.argv[2], N_TRIALS)
+    except KeyboardInterrupt:
+        pass
+    else:
+        new_value = study.best_value
+        _value = study.get_trials()[0].value
+        if new_value > _value:
+            res = study.best_params
+            print(json.dumps(res.update({'new_value': new_value, '_value': _value})))
+    print(json.dumps({}))